+++ lib/arel_extensions/visitors/oracle.rb
@@ -13,6 +13,20 @@ module ArelExtensions
       }
       NUMBER_COMMA_MAPPING = {'en_US' => '.,', 'fr_FR' => ',', 'sv_SE' => ', '}
 
+      def visit_ArelExtensions_Nodes_ByteSize o, collector
+        collector << 'LENGTHB('
+        collector = visit o.expr.coalesce(''), collector
+        collector << ')'
+        collector
+      end
+
+      def visit_ArelExtensions_Nodes_CharLength o, collector
+        collector << 'LENGTH('
+        collector = visit o.expr.coalesce(''), collector
+        collector << ')'
+        collector
+      end
+
       def visit_ArelExtensions_Nodes_Log10 o, collector
         collector << 'LOG('
         o.expressions.each_with_index { |arg, i|